Moved matchers and stubs to their own directories
