+++ ksql-functional-tests/src/main/java/io/confluent/ksql/engine/FakeInsertValuesExecutor.java
@@ -18,8 +18,8 @@ package io.confluent.ksql.engine;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.annotations.VisibleForTesting;
 import io.confluent.ksql.test.serde.string.StringSerdeSupplier;
-import io.confluent.ksql.test.tools.FakeKafkaRecord;
-import io.confluent.ksql.test.tools.FakeKafkaService;
+import io.confluent.ksql.test.tools.stubs.FakeKafkaRecord;
+import io.confluent.ksql.test.tools.stubs.FakeKafkaService;
 import io.confluent.ksql.test.tools.Record;
 import io.confluent.ksql.test.tools.Topic;
 import io.confluent.ksql.test.tools.exceptions.InvalidFieldException;
