+++ ksql-functional-tests/src/main/java/io/confluent/ksql/test/model/KeyFieldNode.java
@@ -22,7 +22,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.confluent.ksql.metastore.model.KeyField;
 import io.confluent.ksql.schema.ksql.types.SqlType;
 import io.confluent.ksql.schema.ksql.types.SqlTypes;
-import io.confluent.ksql.test.model.MetaStoreMatchers.KeyFieldMatchers;
+import io.confluent.ksql.test.model.matchers.MetaStoreMatchers.KeyFieldMatchers;
 import io.confluent.ksql.test.serde.KeyFieldDeserializer;
 import java.util.Objects;
 import java.util.Optional;
