+++ ksql-functional-tests/src/main/java/io/confluent/ksql/test/model/matchers/MetaStoreMatchers.java
@@ -13,7 +13,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.ksql.test.model;
+package io.confluent.ksql.test.model.matchers;
 
 import static org.hamcrest.Matchers.is;
