+++ ksql-functional-tests/src/main/java/io/confluent/ksql/test/tools/TestExecutor.java
@@ -37,6 +37,10 @@ import io.confluent.ksql.services.DefaultServiceContext;
 import io.confluent.ksql.services.ServiceContext;
 import io.confluent.ksql.test.serde.avro.AvroSerdeSupplier;
 import io.confluent.ksql.test.serde.avro.ValueSpecAvroSerdeSupplier;
+import io.confluent.ksql.test.tools.stubs.FakeKafkaClientSupplier;
+import io.confluent.ksql.test.tools.stubs.FakeKafkaRecord;
+import io.confluent.ksql.test.tools.stubs.FakeKafkaService;
+import io.confluent.ksql.test.tools.stubs.FakeKafkaTopicClient;
 import io.confluent.ksql.util.KsqlConfig;
 import io.confluent.ksql.util.KsqlException;
 import java.io.Closeable;
