+++ ksql-functional-tests/src/main/java/io/confluent/ksql/test/tools/stubs/FakeKafkaClientSupplier.java
@@ -13,7 +13,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.ksql.test.tools;
+package io.confluent.ksql.test.tools.stubs;
 
 import java.util.Collections;
 import java.util.Map;
