+++ ksql-functional-tests/src/test/java/io/confluent/ksql/engine/FakeInsertValuesExecutorTest.java
@@ -26,8 +26,8 @@ import com.google.common.collect.ImmutableMap;
 import io.confluent.ksql.engine.FakeInsertValuesExecutor.FakeProduer;
 import io.confluent.ksql.test.serde.avro.AvroSerdeSupplier;
 import io.confluent.ksql.test.serde.string.StringSerdeSupplier;
-import io.confluent.ksql.test.tools.FakeKafkaRecord;
-import io.confluent.ksql.test.tools.FakeKafkaService;
+import io.confluent.ksql.test.tools.stubs.FakeKafkaRecord;
+import io.confluent.ksql.test.tools.stubs.FakeKafkaService;
 import io.confluent.ksql.test.tools.Record;
 import io.confluent.ksql.test.tools.Topic;
 import java.nio.charset.StandardCharsets;
